+++ tests/Composer/Factory/CustomRepositoriesTest.php
@@ -6,16 +6,16 @@ use Dex\Composer\PlugAndPlay\Tests\FactoryTestCase;
 
 class CustomRepositoriesTest extends FactoryTestCase
 {
-    protected function path(): string
+    protected function fixture(): string
     {
-        return __DIR__ . '/../../../fixtures/custom-repositories/';
+        return 'custom-repositories';
     }
 
     public function testFactory(): void
     {
         $this->factory();
 
-        $this->assertOutput('Plugged: dex/fake');
+        $this->assertOutputContains('Plugged: dex/fake');
         $this->assertGeneratedJsonEquals([
             'config' => [
                 'allow-plugins' => true,
